requirements.txt
asgiref==3.2.3
bottle==0.12.19   
cachetools==4.2.0 
certifi==2020.12.5
chardet==3.0.4    
colorama==0.4.4
dj-database-url==0.5.0
Django==3.0
django-crispy-forms==1.8.1
django-environ==0.4.5
django-heroku==0.3.1
eth-hash==0.2.0
eth-typing==2.2.1
google-api-core==1.24.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gunicorn==20.0.4
honcho==1.0.1
httplib2==0.18.1
idna==2.10
oauthlib==3.1.0
Pillow==6.2.1
protobuf==3.14.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.7
pytz==2019.3
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
toolz==0.10.0
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.0.1